advanced relationships with entity framework

+++ Services/CharacterService .cs
@@ -6,9 +6,8 @@ using dotNETCoreWebAPI.Dtos.Character;
 using AutoMapper;
 using dotNETCoreWebAPI.Data;
 using Microsoft.EntityFrameworkCore;
-
-
-
+using Microsoft.AspNetCore.Http;
+using System.Security.Claims;
 
 namespace dotNETCoreWebAPI.Services
 {
@@ -16,13 +15,17 @@ namespace dotNETCoreWebAPI.Services
     {
 
         private readonly IMapper _mapper;
+        private readonly IHttpContextAccessor _httpContext;
         private readonly DataContext _context;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
           _mapper=mapper;
           _context= context;  
+          _httpContext = httpContextAccessor;
         }
+
+        private int GetUserId() => int.Parse(_httpContext.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
          private static List<Character> characters = new List<Character>() {
             new Character(),
@@ -32,16 +35,17 @@ namespace dotNETCoreWebAPI.Services
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character newCharacter = _mapper.Map<Character>(newChar);
+            newCharacter.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
              await _context.characters.AddAsync(newCharacter);
              await _context.SaveChangesAsync();
-             serviceResponse.Data= await _context.characters.Select(c=>_mapper.Map<GetCharacterDto>(c)).ToListAsync();
+             serviceResponse.Data= await _context.characters.Where(c=> c.User.Id==GetUserId()).Select(c=>_mapper.Map<GetCharacterDto>(c)).ToListAsync();
         return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            List<Character> Dbcharacters = await _context.characters.Where(c=> c.User.Id==userId).ToListAsync();
+            List<Character> Dbcharacters = await _context.characters.Where(c=> c.User.Id==GetUserId()).ToListAsync();
             serviceResponse.Data=Dbcharacters.Select(c=>_mapper.Map<GetCharacterDto>(c)).ToList();
             return serviceResponse;
         }
@@ -49,7 +53,7 @@ namespace dotNETCoreWebAPI.Services
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
-            Character DbCharacter =  await _context.characters.FirstOrDefaultAsync(c => c.Id==id);
+            Character DbCharacter =  await _context.characters.FirstOrDefaultAsync(c => c.Id==id && c.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(DbCharacter);
             return serviceResponse;
         }
@@ -59,7 +63,9 @@ namespace dotNETCoreWebAPI.Services
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
          try
          {
-            Character character = await _context.characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+            Character character = await _context.characters.Include(c => c.User).FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+            if(character.User.Id == GetUserId())
+            {
             character.Name = updatedCharacter.Name;
             character.Class = updatedCharacter.Class;
             character.Defence = updatedCharacter.Defense;
@@ -69,6 +75,12 @@ namespace dotNETCoreWebAPI.Services
             _context.characters.Update(character);
             await _context.SaveChangesAsync();
             serviceResponse.Data=_mapper.Map<GetCharacterDto>(character);
+            }
+            else
+            {
+                serviceResponse.Success= false;
+                serviceResponse.Message="Character not found";
+            }
          }
          catch (System.Exception ex)
          {
@@ -88,11 +100,18 @@ namespace dotNETCoreWebAPI.Services
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
            try
            {
-            Character character = await _context.characters.FirstAsync(c => c.Id == id);
-             _context.characters.Remove(character);
+            Character character = await _context.characters.FirstOrDefaultAsync(c => c.Id == id &&  c.User.Id == GetUserId());
+            if(character != null)
+            {
+            _context.characters.Remove(character);
              await _context.SaveChangesAsync();
             serviceResponse.Data = await _context.characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
-               
+            }
+            else
+            {
+                serviceResponse.Success=false;
+                serviceResponse.Message="Character not found";
+            }
            }
            catch (System.Exception ex)
            {
